Docs maintenance (#39)
* Pin versions of sphinx and optional extensions

As recommended for reproducible builds on readthedocs:
https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
and to ensure that our docs build works with readthedocs changes to default
project dependencies:
https://blog.readthedocs.com/defaulting-latest-build-tools/

* Update ReadTheDocs build config

Upgraded the OS to Ubuntu 24.04 and updated the Python toolchain to
mambaforge-23.11 in the .readthedocs.yaml file. This ensures compatibility with
the latest environment settings.

* Add explicit Sphinx configuration for readthedocs

Updated `.readthedocs.yaml` to define Sphinx builder settings, including the
builder type, configuration file path, and `fail_on_warning` option. This is
necessary due to readthedocs deprecating projects without explicit builder
configuration that comes into effect on 20-Jan-2025.

* Add sphinx-notfound-page extension to docs build

Ensures that static assets load in the event of a 404 for a better UX.
Facilitates future customization of 404 page.